Use watch'es persistent storage
Draw screen based on cached data, don't wait for phone connection to become available.

+++ src/c/util.h
@@ -8,6 +8,5 @@ int my_abs(int in);
 double round(double number);
 double myatof (const char *p);
 void ftoa(char* str, double val, int precision);
-//float tm_to_solar_time(struct tm in_tm, int16_t in_solar_offset);
-float hm_to_time(Tuple *h, Tuple *m);
+uint32_t hm_to_time(Tuple *h, Tuple *m);
 char *translate_AppMessageResult(AppMessageResult result);
